Sample output format for Acronymsdown documents

Format for converting R Markdown documents to (pandoc) markdown documents. This format is primarily used for testing purposes.

Usage

md_format(...)

Arguments

...

Additional arguments, which will be passed to the base rmarkdown::md_document() format.

Value

A rmarkdown output format, which can be used in rmarkdown::render().

Details

This output format was meant for testing the Acronymsdown package. You may use it in your documents, but be warned that there are some opinionated choices. In particular, this format:

ensures that headings use the ATX syntax (e.g., # Title, ## Subtitle) ;

preserves line wrappings ;

forces the pandoc variant of markdown (instead of the default, strict markdown) ;

includes the acronymsdown Lua Filter.

These choices make sense in order to facilitate the automated testing of Acronymsdown, by ensuring a reliable (and reproducible) result, even for older versions of pandoc.